requirements.txt
streamlit
langchain
langgraph
chromadb
pydantic
python-dotenv
pypdf
tiktoken
openai
langchain-openai
langchain-community
langchain-text-splitters
langchain-chroma
